moving states tester
